purego: fix cgo fallback build for linux/mips
Package purego failed to compile under CGO_ENABLED=1 for the Linux
GOARCHs that lack a per-architecture struct file (mips, mipsle,
mips64, mips64le). func.go and syscall.go reference helpers
(addStruct, getStruct, shouldBundleStackArgs, collectStackArgs,
bundleStackArgs, syscall_syscallN) that are only defined for the
natively supported architectures, so the documented integer-only
cgo fallback was broken for these targets.

Add struct_other.go, providing those struct helpers as panicking
stubs for architectures that have no dedicated struct_<arch>.go,
and add the syscall_syscallN stub to syscall_cgo_linux.go.
struct_386.go was already a pure stub, so it is removed and 386
now shares struct_other.go. Structs and floats remain unsupported,
exactly as the README's "Cgo Fallback" section documents.

Add a CI step that cross-compiles linux/mips with CGO_ENABLED=1 to
guard against this regression.

Closes #460
